+++ src/Command/Rendering/RendersRuntimeDto.php
@@ -318,6 +318,7 @@ trait RendersRuntimeDto
             'namespace' => $namespace,
             'imports' => $useStatements,
             'className' => $className,
+            'classDeprecated' => $this->deprecatedByClass[$className] ?? false,
             'sourceEndpoint' => $this->endpointByClass[$className] ?? null,
             'sourceSpecLink' => $this->resolveSpecLink($className),
             'sourceRelated' => $this->relatedByClass[$className] ?? null,